requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0
langchain-core>=0.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
openai>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pypdf>=4.0.0
docx2txt>=0.8
pandas>=2.0.0
openpyxl>=3.1.0
